+++ packages/core/avatar/src/types.js
@@ -1,6 +1,8 @@
 // @flow
+
 import type { Node, ComponentType } from 'react';
 import type { AnalyticsEventInterface } from '@atlaskit/analytics-next';
+import type { ThemeProps } from './theme';
 
 export type AppearanceType = 'circle' | 'square';
 export type PresenceType = ('online' | 'busy' | 'focus' | 'offline') | Node;
@@ -73,6 +75,9 @@ export type AvatarPropTypesBase = {
   /** You should not be accessing this prop under any circumstances. It is
    provided by @atlaskit/analytics-next. */
   createAnalyticsEvent?: any,
+
+  /** The theme that should be applied to the avatar. */
+  theme?: ThemeProps => ThemeProps,
 };
 
 export type AvatarPropTypes = AvatarPropTypesBase & {